create climate.py from init
